pyproject.toml
[tool.poetry]
name = "pciw.py"
version = "0.10.0.dev1"
license = "MIT"
readme = "README.md"
description = "Module for obtaining system information."
authors = ["Romanin <semina054@gmail.com>"]
include = ["changelog.md"]
packages = [{include = "pciw"}]
repository = "https://github.com/romanin-rf/pciw.py"
keywords = ["pciw", "pciw.py", "info", "system", "os", "bios", "videocard", "motherboard", "ram"]
classifiers = [
    "Topic :: System",
    "Topic :: System :: Hardware",
    "Topic :: System :: Operating System",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix"
]

[tool.poetry.dependencies]
python = "^3.9"
screeninfo = "^0.8.1"
py-cpuinfo = "^9.0.0"
python-dateutil = "^2.8.2"
rich = "^13.6.0"
pynvml = "^11.5.0"
elevate = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
